requirements.txt
altgraph==0.17.3
aniso8601==9.0.1
APScheduler==3.10.4
argcomplete==1.10.3
ascii==3.6
Babel==2.12.1
bcrypt==4.0.1
beautifulsoup4==4.8.2
capstone==5.0.0rc2
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
colored-traceback==0.3.0
compressed-rtf==1.0.6
contourpy==1.2.0
crypto==1.4.1
cryptography==40.0.2
cycler==0.12.1
dnspython==2.3.0
docx==0.2.4
docx2txt==0.8
docxcompose==1.4.0
docxtpl==0.16.6
ebcdic==1.1.1
extract-msg==0.28.7
Flask==2.2.3
Flask-Login==0.6.2
Flask-RESTful==0.3.10
Flask-WTF==1.1.1
fonttools==4.47.2
future==0.18.2
gevent==23.9.1
greenlet==2.0.2
idna==3.4
IMAPClient==2.1.0
intervaltree==3.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jwt==1.3.1
kiwisolver==1.4.5
lxml==4.9.2
Mako==1.2.4
MarkupSafe==2.1.2
matplotlib==3.8.2
MouseInfo==0.1.3
Naked==0.1.32
numpy==1.26.3
olefile==0.47
packaging==23.1
paramiko==3.1.0
pdfminer.six==20191110
pefile==2022.5.30
Pillow==9.2.0
plumbum==1.8.1
psutil==5.9.4
pwn==1.0
pwntools==4.10.0
PyAutoGUI==0.9.54
pycparser==2.21
pycryptodome==3.20.0
pyelftools==0.29
pygame==2.1.2
PyGetWindow==0.0.9
Pygments==2.15.1
pyinstaller==5.6.1
pyinstaller-hooks-contrib==2022.10
PyJWT==2.7.0
pymongo==4.3.3
PyMsgBox==1.0.9
PyNaCl==1.5.0
pyparsing==3.1.1
PyPDF2==3.0.1
pyperclip==1.8.2
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
PyRect==0.2.0
PyScreeze==0.1.30
pyserial==3.5
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==0.8.11
python-pptx==0.6.23
pytweening==1.0.7
pytz==2023.3.post1
pyuic5-tool==0.0.1
pywin32==306
pywin32-ctypes==0.2.0
PyYAML==6.0.1
requests==2.28.2
ROPGadget==7.3
rpyc==5.3.1
schedule==1.1.0
shellescape==3.8.1
six==1.12.0
sortedcontainers==2.4.0
soupsieve==2.5
SpeechRecognition==3.8.1
SQLAlchemy==2.0.4
SQLAlchemy-serializer==1.4.1
textract==1.6.5
typing_extensions==4.5.0
tzdata==2023.4
tzlocal==5.2
unicorn==2.0.1.post1
urllib3==1.26.14
waitress==2.1.2
Werkzeug==2.2.2
WTForms==3.0.1
xlrd==1.2.0
XlsxWriter==3.1.9
zope.event==5.0
zope.interface==6.1